change version, fix selector issue, fix circular dependency
