+++ packages/devtools-extension/package.json
@@ -6,7 +6,7 @@
     "cra-template": "1.1.2",
     "lodash": "^4.17.21",
     "react": "16.13.0",
-    "react-async-states": "0.1.0-beta-14",
+    "react-async-states": "0.1.1-beta-1",
     "react-dom": "16.13.0",
     "react-json-view": "^1.21.3",
     "react-scripts": "4.0.3"
