+++ packages/docs/docs/api/producer-function.md
@@ -269,5 +269,4 @@ The props runners scope is cascaded:
 - and all cascading calls will be granted the same power.
 - If you run from outside the context provider, your producers will only be able
 to run a source object or a bare producer function.
-
 :::
