+++ packages/example/package.json
@@ -9,7 +9,7 @@
     "lodash": "^4.17.21",
     "prop-types": "^15.6.2",
     "react": "16.13.0",
-    "react-async-states": "0.1.0-beta-14",
+    "react-async-states": "0.1.1-beta-1",
     "react-dom": "16.13.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3"
