+++ packages/react-async-states/src/hooks/utils/readConfig.ts
@@ -63,12 +63,3 @@ export function readUserConfiguration<T, E>(
     userConfig
   );
 }
-
-const defaultAnonymousPrefix = "anonymous-async-state-";
-export const nextKey: () => string = (function autoKey() {
-  let key = 0;
-  return function incrementAndGet() {
-    key += 1;
-    return `${defaultAnonymousPrefix}${key}`;
-  }
-}());
